// apps/derrops-portal/src/client/derrops-cloud/models/service.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Derrops API
 * Internal API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface Service {
    /**
     * Primary key in database
     */
    'id': string;
    /**
     * User ID who owns this service
     */
    'user_id': string;
    /**
     * Service name
     */
    'name': string;
    /**
     * Service endpoint URL
     */
    'endpoint': string;
    /**
     * OpenAPI document URL
     */
    'openapi_doc_url'?: string | null;
    /**
     * OpenAPI document content (stored as text)
     */
    'openapi_doc_content'?: string | null;
    /**
     * Service availability percentage
     */
    'availability'?: number | null;
    /**
     * Average response time in milliseconds
     */
    'response_time'?: number | null;
    /**
     * Date when the service was created
     */
    'created_at': string;
    /**
     * Date when the service was last updated
     */
    'updated_at': string;
}